LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/module/import-meta/import-meta-object.any.serviceworker-module-expected.txt


PASS import.meta is an object
PASS import.meta is extensible
PASS import.meta's properties are writable, configurable, and enumerable
